PMMA-83 hook action for capture, refund and chargeback added, invoice code for capture added

+++ app/code/community/Paymill/Paymill/Model/Method/MethodModelAbstract.php
@@ -294,7 +294,7 @@ abstract class Paymill_Paymill_Model_Method_MethodModelAbstract extends Mage_Pay
     private function _getPrenotificationDate($order)
     {
         $dateTime = new DateTime($order->getCreatedAt());
-        $dateTime->modify('+' . Mage::helper('paymill/optionHelper')->getPrenotificationDays() . ' day');
+        $dateTime->modify('+' . (int) Mage::helper('paymill/optionHelper')->getPrenotificationDays() . ' day');
         $date = Mage::app()->getLocale()->storeDate(
             $order->getStore(), 
             Varien_Date::toTimestamp($dateTime->format('Y-m-d H:i:s')),
